Correctly fix bug #2903, failure to rebuild when linker options change.
The failure to rebuild when linker options change was introduced in
abded0675444, "Add library version support to Shared Lib builder",
between 2.2.0 and 2.3.0.

Turning ShlinkAction into a FunctionAction instead of a CommandAction
made it stop depending on $SHLINKCOM.  Normally, a CommandAction calls
get_presig (Action.py:815) which removes $( ... $) and uses the rest
of that string (recursively fully substituted) as the contents to
hash.

FunctionActions only look at the body of the function, so that removed
the dependency on $SHLINKCOM altogether.  Adding it back in the
varlist does this (Action.py:443):
        for v in vl:
            result.append(env.subst('${'+v+'}'))
so it deep-substitutes the whole thing, including all the $( ... $)
parts.  (All varlist vars do this, not just in FunctionActions.)

What we really want is to depend on the value of env['SHLINKCOM']
in the way CommandActions do, i.e. without the $( ... $)
parts, definitely not the fully substituted version of it.

I'm pretty sure the ignored $(...$) parts should not ever be
included in the signature, so this change updates the varlist handling
code to work the way CommandActions always have.

This change also renames the test files to use the correct bug
number and updates the test.